requirements.txt
wheel==0.30.0
bottle==0.12.18
pymodbus==2.4.0
pyserial==3.4
PyYAML==5.3.1
